Add statuses to tasks

+++ src/main/java/com/potion/ISPotion/Controllers/TaskController.java
@@ -1,14 +1,23 @@
 package com.potion.ISPotion.Controllers;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import com.potion.ISPotion.Classes.Role;
 import com.potion.ISPotion.Classes.Task;
+import com.potion.ISPotion.Classes.TaskStatus;
 import com.potion.ISPotion.Classes.User;
 import com.potion.ISPotion.repo.TaskRepository;
 import com.potion.ISPotion.repo.UserRepository;
+import com.potion.ISPotion.utils.AuthUtils;
+import com.potion.ISPotion.utils.TaskService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.annotation.CurrentSecurityContext;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
 
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
 
 @Controller
@@ -20,6 +29,9 @@ public class TaskController {
 
     @Autowired
     private UserRepository userRepository;
+
+    @Autowired
+    private TaskService taskService;
 
     @GetMapping
     public String getAllTasks(Model model) {
@@ -36,7 +48,21 @@ public class TaskController {
     }
 
     @PostMapping("/new")
-    public String createTask(@RequestParam String description, @RequestParam String username) {
+    public String createTask(@CurrentSecurityContext(expression = "authentication")
+                             Authentication authentication,
+                             @RequestParam String description, @RequestParam String username) {
+        Collection<Role> allowedRoles = new HashSet<>(Arrays.asList(
+                Role.ADMIN,
+                Role.DIRECTOR,
+                Role.HEAD,
+                Role.MERLIN,
+                Role.USER
+        ));
+
+        Collection<Role> userRoles = AuthUtils.getRolesByAuthentication(userRepository, authentication);
+        if (!AuthUtils.anyAllowedRole(userRoles, allowedRoles))
+            return "redirect:/home";
+
         Task task = new Task();
         task.setDescription(description);
 
@@ -46,13 +72,74 @@ public class TaskController {
         }
 
         task.setUser(user);
+        task.setStatus(TaskStatus.ASSIGNED);
         taskRepository.save(task);
         return "redirect:/tasks";
     }
 
     @PostMapping("/{taskId}")
-    public String deleteTaskPost(@PathVariable Long taskId) {
+    public String deleteTaskPost(@CurrentSecurityContext(expression = "authentication")
+                                 Authentication authentication,
+                                 @PathVariable Long taskId) {
+        Collection<Role> allowedRoles = new HashSet<>(Arrays.asList(
+                Role.ADMIN,
+                Role.DIRECTOR,
+                Role.HEAD,
+                Role.MERLIN,
+                Role.USER
+        ));
+
+        Collection<Role> userRoles = AuthUtils.getRolesByAuthentication(userRepository, authentication);
+        if (!AuthUtils.anyAllowedRole(userRoles, allowedRoles))
+            return "redirect:/home";
+
         taskRepository.deleteById(taskId);
         return "redirect:/tasks";
     }
+
+    @PostMapping("/{taskId}/changeStatus")
+    public String changeStatus(@PathVariable Long taskId, @RequestParam String newStatus,
+                               @CurrentSecurityContext(expression = "authentication") Authentication authentication) {
+        Task task = taskRepository.findById(taskId).orElseThrow(() -> new RuntimeException("Задание не найдено"));
+
+        if ("STARTED".equals(newStatus) || "SENT_FOR_REVIEW".equals(newStatus)) {
+            task.setStatus(TaskStatus.valueOf(newStatus));
+        } else {
+            // Проверяем роли для остальных статусов
+            Collection<Role> allowedRoles = new HashSet<>(Arrays.asList(
+                    Role.ADMIN,
+                    Role.DIRECTOR,
+                    Role.HEAD,
+                    Role.MERLIN,
+                    Role.USER
+            ));
+
+            Collection<Role> userRoles = AuthUtils.getRolesByAuthentication(userRepository, authentication);
+
+            if (AuthUtils.anyAllowedRole(userRoles, allowedRoles)) {
+                switch (newStatus) {
+                    case "ASSIGNED":
+                        task.setStatus(TaskStatus.ASSIGNED);
+                        break;
+                    case "REVIEW_STARTED":
+                        task.setStatus(TaskStatus.REVIEW_STARTED);
+                        break;
+                    case "SENT_FOR_REWORK":
+                        task.setStatus(TaskStatus.SENT_FOR_REWORK);
+                        break;
+                    case "ACCEPT_COMPLETED":
+                        task.setStatus(TaskStatus.ACCEPT_COMPLETED);
+                        break;
+                    default:
+                        return "redirect:/tasks";
+                }
+            } else {
+                return "redirect:/tasks";
+            }
+        }
+
+        taskRepository.save(task);
+        return "redirect:/tasks";
+    }
+
 }
